requirements.txt
quart
python-magic
pillow-simd
hypercorn